Refactor UPIComponent to make it more readable
